+++ src/kibana/netmon_libs/custom_modules/save_rule/modal/save-rule.html
