+++ src/kibana/netmon_libs/custom_modules/save_rule/services/kbnIndexService.js
