+++ src/kibana/netmon_libs/custom_modules/save_rule/services/searchAuditor.js
