+++ src/kibana/plugins/dashboard/index.html
@@ -54,7 +54,7 @@
           </button>
           <button type="button" class="btn btn-default"
           style="margin-left:3px"
-          ng-disabled="queryInput.$invalid" aria-label="Save as Rule" ng-click="openSaveRuleModal(state.query)">
+          ng-disabled="queryInput.$invalid" aria-label="Save as Rule" ng-click="saveRuleManager.convertAndOpenModal(state)">
            <span aria-hidden="true">Save Rule</span>
           </button>
         </div>
